Call to wallet_addEthereumChain() is now performed by EIP6963Agent in place of WalletCient_Ethereum.
